requirements.txt
anyio==4.6.2.post1
beautifulsoup4==4.12.3
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
cloudevents==1.9.0
cryptography==43.0.3
deprecation==2.1.0
exceptiongroup==1.2.2
fastapi==0.85.1
firebase-admin==6.6.0
firebase-functions==0.1.2
Flask==2.3.3
Flask-Cors==5.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-events==0.13.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.2
PyJWT==2.9.0
pyparsing==3.2.0
pypdfium2==4.30.0
python-dotenv==1.0.1
python-multipart==0.0.12
python-pptx==1.0.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.28.1
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
starlette==0.20.4
tqdm==4.67.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.19.0
#functions-framework==3.0.0
#watchdog==1.0.2
Werkzeug==3.1.3
XlsxWriter==3.2.0
